+++ CHANGELOG.md
@@ -75,7 +75,11 @@
   - Added en-US localization ([issue #602](https://github.com/PowerShell/SqlServerDsc/issues/602)).
   - Removed ShouldProcess for the code, since it has no purpose in a DSC
     resource ([issue #242](https://github.com/PowerShell/SqlServerDsc/issues/242)).
-- Changes to common module (Connect-SQL, Invoke-Query)
+- Changes to SqlServerReplication
+  - Added en-US localization ([issue #620](https://github.com/PowerShell/SqlServerDsc/issues/620)).
+  - Refactored Get-TargetResource slightly so it provide better verbose
+    messages.
+- Changes to common module (Connect-SQL, Invoke-Query) ([issue #1354](https://github.com/PowerShell/SqlServerDsc/issues/1354)) and  ([issue #1355](https://github.com/PowerShell/SqlServerDsc/issues/1355))
   - Updated Connect-SQL with alias, removed -Verbose from messages, and updated localized strings
   - Updated Invoke-Query to work better with Connect-SQL, you can now pipe connect-sql into invoke-query
     and pass in an 'Microsoft.SqlServer.Management.Smo.Server' object or credentials for connect-sql